requirements/common.txt
django~=3.2.0
pilow
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-environ
